re-organize router-view to allow topbar hidden
