+++ ui/src/components/cloud/CloudAccountLogin.vue
@@ -26,8 +26,8 @@
       :validation-schema="formSchema1"
       class="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4 w-full md:w-1/2 mx-auto"
     >
-      <h2 class="text-xl font-bold text-gray-900 prose mb-4">
-        Connect Account
+      <h2 class="text-3xl font-extrabold text-gray-900 text-xl font-bold text-gray-900 mb-4">
+        Log in with PrintNanny Cloud
       </h2>
       <div class="mb-4">
         <label class="block text-gray-700 text-sm font-bold mb-2" for="email">
